Reactive auth, calendar data encapsulated
