+++ src/lib/store/index.js
@@ -7,6 +7,9 @@ export { user } from './userStore.js'
 export { clickedTaskID, familyTree, openTaskPopup, closeTaskPopup } from './taskPopup.js'
 export { tasksCache, updateCache, cleanupCache } from './tasksCache.js'
 
+export const authUser = writable(null) // reactive version of firebaseAuth.currentUser (= to resultUser from onAuthStateChanged)
+export const initialDataReady = writable(false)
+
 export const allAccounts = writable([])
 export const cals = writable({})
 export const googleEventsByDate = writable({})
