+++ src/routes/[user]/components/Calendar/DayColumn.svelte
@@ -7,11 +7,10 @@
   import { getLocalY } from '$lib/utils/core.js'
   import { DateTime } from 'luxon'
   import { pixelsPerHour, headerHeight, timestampsColumnWidth } from './store.js'
-  import { treesByDate } from './service.js'
   import { timestamps, calSnapInterval, googleEventsByDate } from '$lib/store'
   import { getContext } from 'svelte'
 
-  const { Task } = getContext('app')
+  const { Task, treesByDate } = getContext('app')
   const { activateInput, overrideOptions } = getContext('popover-input')
   const { 
     registerDropzone,
